+++ internal/variable/schema.go
@@ -29,9 +29,12 @@ type LengthLimited interface {
 	// MaxLength returns the maximum permitted length of the native value.
 	MaxLength() (int, bool)
 
-	// ExplainLengthError returns a human-readable description of the length
-	// constraints, for use in an error message.
-	ExplainLengthError() string
+	// LengthDescription returns a human-readable description of the length that
+	// the limit applies to.
+	//
+	// It must produce a gramatical sentence of the form:
+	//  "The value must have <desc> of exactly 5 bytes.
+	LengthDescription() string
 }
 
 // SchemaError indicates that a value is invalid because it violates its schema.